Create an Utility class
Creation of RangementUtils to simplify Rangement class
