+++ Finish.html
@@ -12,11 +12,11 @@
 <STRONG>Description :</STRONG><BR>
 Programme de gestion de Cave à vins.
 <HR>
-<STRONG>Développements et corrections de la version 5.0 CO (release 2.5.6.8)</STRONG><BR>
+<STRONG>Développements et corrections de la version 5.0 ZE (release 2.5.8.1)</STRONG><BR>
 </FONT>
 <UL>
 <LI>Creer_Tableaux: Il n'était plus possible de créer les tableaux OK
-<LI>Rangement: Refonte de la classe OK
+<LI>Rangement/RangementUtils: Refonte de la classe OK
 <LI>ShowFile: L'affichage de la poubelle faisait planter l'application OK
 <LI>Program: Suppression de la gestion des anciennes versions OK
 <LI>Program: Refonte des calculs des statistiques OK
