+++ src/mycellar/core/MyCellarVersion.java
@@ -9,8 +9,8 @@ import java.io.InputStreamReader;
 
 public class MyCellarVersion {
 
-	public static final String version = "2.5.7.5";
-	public static final String mainVersion = "4.9 CO";
+	public static final String version = "2.5.8.1";
+	public static final String mainVersion = "5.0 ZE";
 	  
 	public static String getLocalVersion() {
 		InputStream stream = MyCellarVersion.class.getClassLoader().getResourceAsStream("MyCellarVersion.txt");
