+++ src/title_en.properties
@@ -293,7 +293,7 @@ Infos292=2 digits year greater than
 Infos293=belong to
 Infos294=Date control
 Infos295=century.
-Infos296=May
+Infos296=July
 Infos297=To copy line number to each part
 Infos298=click OK
 Infos299=bracket with bottles
