+++ src/title_fr.properties
@@ -293,7 +293,7 @@ Infos292=Années de 2 chiffres supérieures à
 Infos293=appartiendront au
 Infos294=Contrôle de date
 Infos295=ème siècle.
-Infos296=Mai
+Infos296=Juillet
 Infos297=Cliquez sur OK du nombre de lignes
 Infos298=pour copier la valeur pour chaque partie
 Infos299=Tranches avec bouteilles
